Split the high five parsing code into its own file so it can be tested independently
